pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.13.3",
    "discord-py>=2.7.1",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
]
